Uploads done in VBA
This version allows the updates to be done without the need for the Stored Procedures. 
The SPs are only needed for auto-detecting the data type. In other words if the user wants to add a new table or column and doesn't want to define the data type, then the SPs are needed.