Removed Role, WH, DB & Schema from login and removed Autogenerate datatypes
Simplified Login. User doesn't have to fill in Role, Warehouse, DB or Schema.
Removed the ability to automitically detect data types. The code is still there if we want to add it back.
Changed the way stages are created. They are now created in the same DB as where the table being updated is. It used to be created in the default database. Since we are not setting a default DB.schema we have to do it in the table's DB.
Now when the user logs in, the default Role and Warehouse are used. This can be chanegd from the Config menu. If the user does not have a default warehouse, they are asked to type one in.