Fixed issue with saving SQL and Saved SQL per worksheet
When users don't have this Excel option checked: Trust access to the VBA-Project Object model
The CodeName for the sheet is empty causing the SQL and Saved SQL to be shared across all worksheets.
This fix will use the worksheet name instean of the code when the Trust option above is not set. This means that if the name of the worksheet changes, the SQL and Saved SQL will be lost.