Issue 21 - If Role = ALL then Role is set to ""
The pseudo role 'ALL' is causing a problem in some Snowflake deployments.  In the connection string, Role = ALL can not be set, so instead it's being set to empty string. If the users default role is ALL then it works fine because the subsequent attempts will empty the role value.